% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.tbl_validation}
\alias{aggregate.tbl_validation}
\title{Aggregate a validation}
\usage{
\method{aggregate}{tbl_validation}(x, by = c("rule", "record"), ...)
}
\arguments{
\item{x}{\code{tbl_validation} object}

\item{by}{either by "rule" or by "record"}

\item{...}{not used}
}
\description{
Create either statistics by rule or record
}
\examples{
income <- data.frame(id = 1:2, age=c(12,35), salary = c(1000,NA))
con <- dbplyr::src_memdb()
tbl_income <- dplyr::copy_to(con, income, overwrite=TRUE)
print(tbl_income)

# Let's define a rule set and confront the table with it:
rules <- validator( is_adult   = age >= 18
                    , has_income = salary > 0
)

# and confront!
cf <- confront(tbl_income, rules, key="id")
aggregate(cf, by = "rule")
aggregate(cf, by = "record")
}
