% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute.tbl_validation}
\alias{compute.tbl_validation}
\title{Store the result of the validation in the db}
\usage{
\method{compute}{tbl_validation}(x, name, ...)
}
\arguments{
\item{x}{\code{tbl_validation} object}

\item{name}{optional, when omitted, a random name is used.}

\item{...}{passed through to \code{compute} on \code{x$query}}
}
\description{
Stores the result of the validation in the db using
the \code{\link[dplyr:compute]{dplyr::compute()}} of the db back-end.
This method changes the \code{tbl_validation} object!
Note that for most back-ends the default setting is
a temporary table with a random name.
}
\seealso{
Other tbl_validation: 
\code{\link{tbl_validation-class}}
}
\concept{tbl_validation}
