% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_tbl_sparse.R
\name{confront_tbl_sparse}
\alias{confront_tbl_sparse}
\title{Create a query object for sparse table validation}
\usage{
confront_tbl_sparse(
  tbl,
  x,
  key = NULL,
  union_all = TRUE,
  ...,
  check_rules = TRUE
)
}
\arguments{
\item{tbl}{tbl object}

\item{x}{\code{validator} object}

\item{key}{character of key column (not working)}

\item{union_all}{if \code{FALSE} each rule is a separate query.}

\item{...}{not}

\item{check_rules}{if \code{TRUE} it is checked which rules 'work' on the db.}
}
\description{
Confront \code{tbl_sql} objects with \code{validator} rules.
}
\concept{confront}
