% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_wide.R
\name{confront_wide}
\alias{confront_wide}
\title{create a table with per record if it abides to the rule.}
\usage{
confront_wide(tbl, x, compute = FALSE, ...)
}
\description{
create a table with per record if it abides to the rule.
}
