% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_sql.R
\name{dump_sql}
\alias{dump_sql}
\title{dump sql statements}
\usage{
dump_sql(x, sql_file = stdout(), sparse = x$sparse, ...)
}
\arguments{
\item{x}{\code{tbl_validation} object}

\item{sql_file}{filename/connection where the sql code should be written to.}

\item{sparse}{not used}

\item{...}{not used}
}
\description{
Write sql statements of a tbl confrontation.
}
