% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_record_based.R
\name{is_record_based}
\alias{is_record_based}
\title{determine the dimensional structure of \code{values} so that it conforms to
\code{validate}. Only used for values \code{simplify}.
Works heuristically: get a subset out of db and executes validate::confront on it.}
\usage{
is_record_based(tbl, x, n = 5)
}
\description{
determine the dimensional structure of \code{values} so that it conforms to
\code{validate}. Only used for values \code{simplify}.
Works heuristically: get a subset out of db and executes validate::confront on it.
}
\keyword{internal}
