% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_works_on_tbl.R
\name{rule_works_on_tbl}
\alias{rule_works_on_tbl}
\title{tests for each rule if it can be executed on the database}
\usage{
rule_works_on_tbl(tbl, x)
}
\arguments{
\item{tbl}{a \code{tbl} object with columns used in \code{x}}

\item{x}{a \code{\link[validate:validator]{validate::validator()}} object}
}
\value{
\code{logical} encoding which validation rules "work" on the database.
}
\description{
tests for each rule if it can be executed on the database
}
