% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_query.R
\name{show_query.tbl_validation}
\alias{show_query.tbl_validation}
\title{Show generated sql code}
\usage{
\method{show_query}{tbl_validation}(x, ...)
}
\arguments{
\item{x}{\code{\link[=tbl_validation]{tbl_validation()}} object, result of a \code{\link[=confront.tbl_sql]{confront.tbl_sql()}}.}

\item{...}{passed through.}
}
\description{
Shows the generated sql code for the validation of the tbl.
}
