% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_validation.R
\docType{class}
\name{tbl_validation-class}
\alias{tbl_validation-class}
\alias{tbl_validation}
\title{Validation object}
\description{
Validation information
}

\seealso{
Other validation: 
\code{\link{confront.tbl_sql}()},
\code{\link{values,tbl_validation-method}}

Other tbl_validation: 
\code{\link{compute.tbl_validation}()}
}
\concept{tbl_validation}
\concept{validation}
