% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatedb-package.R
\docType{package}
\name{validatedb-package}
\alias{validatedb}
\alias{validatedb-package}
\title{validatedb: Validate on a table in a DB, using dbplyr}
\description{
Validate on a table in a DB, using dbplyr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwindj/validatedb}
  \item Report bugs at \url{https://github.com/edwindj/validatedb/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
