% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute.tbl_validation}
\alias{compute.tbl_validation}
\title{Store the validation result in the db}
\usage{
\method{compute}{tbl_validation}(x, name, ...)
}
\arguments{
\item{x}{\code{\link[=tbl_validation]{tbl_validation()}}, result of a \code{confront()} of \code{tbl} with a rule set.}

\item{name}{optional, when omitted, a random name is used.}

\item{...}{passed through to \code{compute} on \code{x$query}}
}
\value{
A \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} object that refers to the computed (temporary)
table in the database. See \code{\link[dplyr:compute]{dplyr::compute()}}.
}
\description{
Stores the validation result in the db using
the \code{\link[dplyr:compute]{dplyr::compute()}} of the db back-end.
This method changes the \code{tbl_validation} object!
Note that for most back-ends the default setting is
a temporary table with a random name.
}
\seealso{
Other tbl_validation: 
\code{\link{tbl_validation-class}}
}
\concept{tbl_validation}
