% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront.R
\name{confront,ANY,validator,ANY-method}
\alias{confront,ANY,validator,ANY-method}
\title{Implementation for database tables.
This method is used for object that implement the \code{tbl_sql} interface used in
\code{dbplyr}.}
\usage{
\S4method{confront}{ANY,validator,ANY}(dat, x, ref, key = NULL, sparse = FALSE, ...)
}
\arguments{
\item{dat}{an object of class `tbl_sql``.}

\item{x}{\code{validator} object}

\item{ref}{reference object (not working)}

\item{key}{character of key column (not working)}

\item{sparse}{\code{logical} should only errors be stored in the db?}

\item{...}{not used}
}
\description{
Implementation for database tables.
This method is used for object that implement the \code{tbl_sql} interface used in
\code{dbplyr}.
}
