% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront.R
\name{confront.tbl_sql}
\alias{confront.tbl_sql}
\title{Confront \code{tbl_sql} objects with \code{validator} rules.}
\usage{
confront.tbl_sql(tbl, x, ref, key = NULL, sparse = FALSE, compute = FALSE, ...)
}
\arguments{
\item{tbl}{tbl object}

\item{x}{\code{validator} object}

\item{ref}{reference object (not working)}

\item{key}{character of key column (not working)}

\item{sparse}{\code{logical} should only errors be stored in the db?}

\item{compute}{\code{logical} if \code{TRUE} the check stores a temporary table in the database.}

\item{...}{not}
}
\description{
Confront \code{tbl_sql} objects with \code{validator} rules.
}
