% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_tbl.R
\name{confront_tbl}
\alias{confront_tbl}
\title{create a table with per record if it abides to the rule.}
\usage{
confront_tbl(tbl, x, key = NULL, ...)
}
\arguments{
\item{tbl}{tbl object}

\item{x}{\code{validator} object}

\item{key}{character of key column (not working)}

\item{...}{not used}
}
\description{
create a table with per record if it abides to the rule.
}
