% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_tbl.R
\name{confront_tbl}
\alias{confront_tbl}
\title{create a table with per record if it abides to the rule.}
\usage{
confront_tbl(tbl, x, key = NULL, ...)
}
\arguments{
\item{tbl}{tbl_sql object, table in a database, retrieved with \code{\link[dplyr:tbl]{dplyr::tbl()}}}

\item{x}{\code{\link[validate:validator]{validate::validator()}} object with validation rules.}

\item{key}{\code{character} with key column name (not working)}

\item{...}{passed through to \code{\link[=compute]{compute()}}, if \code{compute} is \code{TRUE}}
}
\description{
create a table with per record if it abides to the rule.
}
