% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_tbl.R
\name{confront_tbl}
\alias{confront_tbl}
\title{create a table with per record if it abides to the rule.}
\usage{
confront_tbl(tbl, x, key = NULL)
}
\arguments{
\item{tbl}{\code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} table in a database, retrieved with \code{\link[=tbl]{tbl()}}}

\item{x}{\code{\link[validate:validator]{validate::validator()}} object with validation rules.}

\item{key}{\code{character} with key column name.}
}
\value{
a list with needed information, see details.
}
\description{
create a table with per record if it abides to the rule.
}
\details{
The return value of the function is a list with:
\itemize{
\item \verb{$query}: A \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} object that refers to the confrontation query.
\item \verb{$errors}: The validation rules that are not working on the database
\item \verb{$working}: A \code{logical} with which expression are working on the database.
\item \verb{$exprs}: All validation expressions.
\item \verb{$nexprs}: Number of working expression.
}
}
