% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confront_sparse_qry.R
\name{confront_tbl_sparse}
\alias{confront_tbl_sparse}
\title{create a table with only fails}
\usage{
confront_tbl_sparse(
  tbl,
  x,
  compute = FALSE,
  union_all = TRUE,
  ...,
  check_rules = TRUE
)
}
\description{
create a table with only fails
}
