% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R, R/show_query.R, R/tbl.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compute}
\alias{show_query}
\alias{tbl}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{compute}}, \code{\link[dplyr:explain]{show_query}}, \code{\link[dplyr]{tbl}}}
}}

