% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_validation.R
\docType{class}
\name{tbl_validation-class}
\alias{tbl_validation-class}
\alias{tbl_validation}
\title{Validation object}
\description{
Validation information
}

\seealso{
Other confront: 
\code{\link{confront.tbl_sql}()},
\code{\link{confront_tbl_sparse}()},
\code{\link{values,tbl_validation-method}}
}
\concept{confront}
