% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatedb-package.R
\docType{package}
\name{validatedb-package}
\alias{validatedb}
\alias{validatedb-package}
\title{validatedb: Validate Data in a Database using 'validate'}
\description{
Check whether records in a database table are valid using 
   validation rules in R syntax specified with R package 'validate'. 
   R validation checks are automatically translated in SQL using 'dbplyr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-cleaning/validatedb}
  \item Report bugs at \url{https://github.com/data-cleaning/validatedb/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/0000-0002-6580-4718}{ORCID})

Authors:
\itemize{
  \item Mark van der Loo
}

}
\keyword{internal}
