% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values,tbl_validation-method}
\alias{values,tbl_validation-method}
\title{testing 1, 1,2}
\usage{
\S4method{values}{tbl_validation}(x, simplify = FALSE, type = c("list", "tbl"), ...)
}
\description{
testing 1, 1,2
}
